+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   * Support for issue types when creating, editing, and filtering issues.
   * Support sorting merge requests by title.
   * Support requesting retried jobs when requesting pipeline jobs.
+  * Support pipeline status states `Preparing` and `WaitingForResource`.
 
 ## Deprecations